fix: only fetch supported record types from the Mikrotik API (#135)
The problem was that the webhook was sending an API request to the Mikrotik RouterOS API to get a list of all of the configured DNS records. This would cause an issue since Mikrotik does support some record types that external DNS does NOT support.

This PR adds some filtering to only get a list of supported DNS record types back from the RouterOS API.